+++ src/testRunner/activateRunner.ts
@@ -299,7 +299,7 @@ function handleTestTreeUpdates(
                 updateType == FileChangeType.Modified &&
                 item.isFile() &&
                 extname(item.getPath()) == getExtensionForBrunoFiles() &&
-                changedData?.sequenceChanged
+                (changedData?.sequenceChanged || changedData?.tagsChanged)
             ) {
                 /* For directories, no changes are ever registered because renaming a directory is seen as a creation of a new directory with the
                 new name and a deletion of the directory with the old name. */